♻️ Refactor: entrypoint 레이어 DTO 구조 변경
